requirements.txt
apispec==3.3.0
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
click==7.1.1
colorama==0.4.3
coverage==5.0.4
Flask==1.1.1
flask-apispec==0.8.7
gunicorn==20.0.4
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
marshmallow==3.5.1
more-itertools==8.2.0
numpy==1.18.2
packaging==20.3
pandas==1.0.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.4.1
pytest-cov==2.8.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
six==1.14.0
wcwidth==0.1.9
webargs==5.5.3
Werkzeug==1.0.0
wincertstore==0.2
zipp==3.1.0
